Tidy and refactor
