+++ prisma/schema.prisma
@@ -895,7 +895,6 @@ enum activity_type_enum {
   AgilityArena
   ChampionsChallenge
   AerialFishing
-  CrystalImpling
   DriftNet
   MahoganyHomes
   GnomeRestaurant
