+++ src/lib/types/minions.ts
@@ -279,14 +279,6 @@ export interface HunterActivityTaskOptions extends ActivityTaskOptions {
 	usingStaminaPotion: boolean;
 }
 
-export interface CrystalImplingActivityTaskOptions extends ActivityTaskOptions {
-	type: 'CrystalImpling';
-	creatureName: string;
-	quantity: number;
-	usingHuntPotion: boolean;
-	usingStaminaPotion: boolean;
-}
-
 export interface AlchingActivityTaskOptions extends ActivityTaskOptions {
 	type: 'Alching';
 	itemID: number;
@@ -606,5 +598,4 @@ export type ActivityTaskData =
 	| FightCavesActivityTaskOptions
 	| ActivityTaskOptionsWithQuantity
 	| MinigameActivityTaskOptionsWithNoChanges
-	| CutLeapingFishActivityTaskOptions
-	| CrystalImplingActivityTaskOptions;
+	| CutLeapingFishActivityTaskOptions;
